+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### 0.6.0
 
-- Implement HashTools#convert_keys_to_strings and HashTools#convert_keys_to_symbols.
+- Implement HashTools#convert_keys_to_strings and #convert_keys_to_symbols.
 - Implement ObjectTools#dig.
+- Implement StringTools#singular? and #plural?.
 - Implement Toolbox::ConstantMap.
 - Implement Toolbox::Mixin.